Add files via upload
this code is definitely not complete. In fact there are several errors. I listed the couple i noticed on the bottom with the pound sign. I just wanted to complete it before midnight because you had already given me an extension. I will gladly take the grade the submitted piece of work deserves, but i would love if you could go over it with me and perhaps we could find the errors together whenever you are free. Thanks.